analyze: convert examples to a proper test suite
