+++ c2rust-analyze/tests/filecheck/alias1.rs
@@ -1,0 +1,40 @@
+use std::ptr;
+
+// CHECK-LABEL: final labeling for "alias1_good"
+pub unsafe fn alias1_good() {
+    // CHECK-DAG: ([[#@LINE+1]]: mut x): addr_of = READ | WRITE | UNIQUE,
+    let mut x = 0;
+    // CHECK-DAG: ([[#@LINE+1]]: p): {{.*}}type = UNIQUE#
+    let p = ptr::addr_of_mut!(x);
+    // CHECK-DAG: ([[#@LINE+1]]: q): {{.*}}type = READ | WRITE | UNIQUE#
+    let q = ptr::addr_of_mut!(x);
+    *q = 1;
+}
+
+// CHECK-LABEL: final labeling for "alias1_bad"
+pub unsafe fn alias1_bad() {
+    // CHECK-DAG: ([[#@LINE+1]]: mut x): addr_of = READ | WRITE,
+    let mut x = 0;
+    // CHECK-DAG: ([[#@LINE+1]]: p): {{.*}}type = READ | WRITE#
+    let p = ptr::addr_of_mut!(x);
+    // CHECK-DAG: ([[#@LINE+1]]: q): {{.*}}type = (empty)#
+    let q = ptr::addr_of_mut!(x);
+    *p = 1;
+}
+
+
+#[cfg(not(compiling_for_test))]
+pub fn safe_alias1_good() {
+    let mut x = 0;
+    let p = &mut x;
+    let q = &mut x;
+    *q = 1;
+}
+
+#[cfg(not(compiling_for_test))]
+pub fn safe_alias1_bad() {
+    let mut x = 0;
+    let p = &mut x;
+    let q = &mut x;
+    *p = 1;
+}
